line/linepython/requirements.txt
Flask==1.1.2
line-bot-sdk==1.16.0
requests==2.24.0
Pillow==7.1.2
opencv-python